SASS-1061 integration tests and refactor
